requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.6.0
idna==3.10
multidict==6.4.3
praw==7.8.1
prawcore==2.4.0
propcache==0.3.1
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.13.2
update-checker==0.18.0
urllib3==2.4.0
websocket-client==1.8.0
yarl==1.20.0
